// src/types/jobs/Info/jd.types.ts
export type JobDescription = {
  jobTitle: string
  location:string
  employementType: string
  compensation: string
  workingdays:number
  workinghours:number
  resorurces:number
  rotation: string
  visa: string
  position_start_date: string
  position_est_end_date: string
  projectduration: string
  overview: string
  responsibilties: string[]
  requirements: string[]
  qualification: string[]
 
}

export type AdvertisedPositionType = {
  nowHiring: string
  location:string
  employementType: string
  startDate: string
  overview: string
  responsibilties: string[]
  requirements: string[]
 
}